move layout components to layout-components folder
