+++ src/components/layout-components/newsGraphContainer.tsx
