+++ src/components/layout-components/newsGraphLayout.tsx
