+++ src/components/layout-components/priceSearchLayout.tsx
